pyproject.toml
[project]
name = "tally"
version = "0.1.0"
description = "Let AI classify your transactions - LLM-powered spending categorization"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "David Fowler"}
]
keywords = ["budget", "finance", "llm", "ai", "categorization", "transactions"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
tally = "tally.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-playwright>=0.4.0",
]
build = [
    "pyinstaller>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tally"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "playwright: tests that require Playwright browser automation",
]
